reworked recent page
